fixed sloppy type casting which was making admiral hopper unhappy
git-svn-id: https://anag-repo.lbl.gov/svn/BISICLES/public/trunk@1833 4b1697f3-9003-4f1e-b6a8-e72787e6a227

+++ code/test/testFortranInterfaceIBC.cpp
@@ -195,7 +195,7 @@ testFortranInterfaceIBC()
   dx /= numCells;
 
   //int maxBoxSize = 8;
-  int numSplit = sqrt(numProc());
+  int numSplit = int(sqrt(float(numProc())));
 
   int maxBoxSize = 32/numSplit;
